+++ examples/basic/pages/index.js
@@ -2,15 +2,18 @@ import React from 'react'
 import Link from 'next/link'
 import fetch from '../libs/fetch'
 
-import { useQuery } from 'react-query'
+import { useQuery, useIsFetching } from 'react-query'
 
 export default () => {
   const { data, isLoading, isFetching } = useQuery('projects', () =>
     fetch('/api/data')
   )
 
+  const isGlobalFetching = useIsFetching()
+
   return (
     <div style={{ textAlign: 'center' }}>
+      {isGlobalFetching ? 'fetching...' : null}
       <h1>Trending Projects</h1>
       <div>
         {isLoading ? (